feat(app): Only display instrument settings for selected robot
closes #2362
